+++ database/migrations/2021_04_03_034201_create_recuperateurs_table.php
@@ -19,7 +19,7 @@ class CreateRecuperateursTable extends Migration
             $table->string('name');
             $table->string('lastname');
             $table->string('email')->unique();
-            $table->integer('phonenumber');
+            $table->bigInteger('phonenumber');
             $table->timestamps();
         });
     }
